fix: ensure initial plot is generated only once
